% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem_module.R
\name{aggregate_extracted_problems}
\alias{aggregate_extracted_problems}
\title{Perform the last wrangling before plotting extreme problems}
\usage{
aggregate_extracted_problems(joined_user_problems, extreme_problems,
  question_choices)
}
\arguments{
\item{joined_user_problems}{The result of `join_users_problems`}

\item{extreme_problems}{The result of `get_extreme_summarised_score`}

\item{question_choices}{The result of `filter_extreme_problem_choices`}
}
\value{
A dataframe with `problem,` `choice` and percent `correct`
}
\description{
Calculates the percent of students selected each option and determines if that option is correct.
}
