% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem_module.R
\name{aggregate_melted_problems}
\alias{aggregate_melted_problems}
\title{Aggregate questions before plotting}
\usage{
aggregate_melted_problems(lookup_table, joined_user_problems, extreme_problems)
}
\arguments{
\item{lookup_table}{The result of `create_question_lookup_from_json`}

\item{joined_user_problems}{The result of `join_users_problems`}

\item{extreme_problems}{The result of `get_extreme_summarised_score`}
}
\value{
A dataframe ready to be plotted by `plot_aggregated_problems`
}
\description{
Aggregate questions before plotting
}
