% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forum_server.R
\name{calculate_forum_searches}
\alias{calculate_forum_searches}
\title{Calculates the number of unique users who searched for each search query, given the selected filters.}
\usage{
calculate_forum_searches(forum_searches = wrangled_forum_searches,
  activity_level, gender, registration_status, category)
}
\arguments{
\item{forum_searches}{The wrangled forum searches dataframe.}

\item{activity_level}{The activity level of the students.}

\item{gender}{The gender of the students.}

\item{registration_status}{The registration status of the students.}

\item{category}{The forum category.}
}
\value{
\code{forum_searches} The filtered forum searches dataframe.
}
\details{

}
\examples{
calculate_forum_searches(wrangled_forum_searches, "under_30_min", "female", "verified", "Part 3")
}
