% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem_module.R
\name{calculate_percent_correct_tbl}
\alias{calculate_percent_correct_tbl}
\title{Calculate the percent of correct answers}
\usage{
calculate_percent_correct_tbl(joined_scores, number_correct)
}
\arguments{
\item{joined_scores}{`problems_tbl` and `lookup_table` joined by `id` and
`item_response`}

\item{number_correct}{the result of `tally_correct_answer`}
}
\value{
A dataframe where each row is a row and the columns show the
  percentage and absolute number of students that got the problem correct.
}
\description{
Each row represents a problem.
}
