% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_wrangling.R
\name{check_integrity}
\alias{check_integrity}
\title{Checks to make sure start and end data passes sanity checks}
\usage{
check_integrity(start, end, max_stop_position)
}
\arguments{
\item{start}{Time into video that the user has started watching the video}

\item{end}{Time into the video that the user has stopped watching the video}

\item{max_stop_position}{Length of the video being watched}
}
\value{
\code{integrity}: Boolean of whether or not the data passes integrity 
  checks
}
\description{
Returns a boolean of whether or not start and end data makes 
  sense. This checks for NA values, end times that are passed the maximum 
  length of the video, and extremely long and short watch durations.
  The threshold for watch durations can be adjusted in the global constants: 
  \code{MIN_DURATION} and \code{MAX_DURATION}
}
\examples{
check_integrity(start, end, max_stop_position)
}
