% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_engagement_wrangling.R
\name{clean_engagement_data}
\alias{clean_engagement_data}
\title{Clean engagement data into tidy format}
\usage{
clean_engagement_data(dirt_engagement_data)
}
\arguments{
\item{dirt_engagement_data}{tower_engage_dirt dataframe object}
}
\value{
dirt_engagement_data
}
\description{
Clean engagement data into tidy format
}
\examples{
clean_engagement_data(dirt_engagement_data = obtain_dirt_engagement_data)
}
