% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem_module.R
\name{clean_multiple_choice}
\alias{clean_multiple_choice}
\title{Clean a demographic multiple choice CSV}
\usage{
clean_multiple_choice(raw_csv)
}
\arguments{
\item{raw_csv}{A dataframe from the read_csv}
}
\value{
A dataframe with the same dimensions as the CSV
}
\description{
This function cleans a CSV retrieved from the `demographic_multiple_choice`
SQL script. It removes any non-multiple choice problems.
}
