% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_server.R
\name{convert_loe}
\alias{convert_loe}
\title{Converts level of education code to explicit string}
\usage{
convert_loe(LoE)
}
\arguments{
\item{LoE}{String value of level of education code}
}
\value{
Returns explicit string value of level of education
}
\description{
Converts level of education code to explicit string
}
