% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forum_server.R
\name{count_authors}
\alias{count_authors}
\title{Count the number of authors for each subcategory in an input forum.}
\usage{
count_authors(input_forum, wrangled_forum_elements = wrangled_forum_elements)
}
\arguments{
\item{input_forum}{The input dataframe containing a row for each post in the forum.}

\item{wrangled_forum_elements}{A dataframe containing information about each of the forum subcategories.}
}
\value{
\code{author_counts} A dataframe containing the unique author counts for each subcategory in the forum.
}
\examples{
count_authors(wrangled_forum_posts)
}
