% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forum_server.R
\name{count_posts}
\alias{count_posts}
\title{Count the number of posts for each subcategory in an input forum.}
\usage{
count_posts(input_forum, wrangled_forum_elements)
}
\arguments{
\item{input_forum}{The input dataframe containing a row for each post in the forum.}

\item{wrangled_forum_elements}{A dataframe containing information about each of the forum subcategories.}
}
\value{
\code{post_counts} A dataframe containing the post counts for each subcategory in the forum.
}
\description{
Count the number of posts for each subcategory in an input forum.
}
\examples{
count_posts(wrangled_forum_posts)
}
