% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_page_server.R
\name{creat_link_table}
\alias{creat_link_table}
\title{create the link summary table : external hyperlink and number of click}
\usage{
creat_link_table(link_summary)
}
\value{
link_df
}
\description{
create the link summary table : external hyperlink and number of click
}
\examples{
creat_link_table(link_summary = reactive_link_dat())
}
