% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_page_server.R
\name{creat_page_table}
\alias{creat_page_table}
\title{Create a page summary table to show page (name, clickable hyperlink) and its 
   page view sorted by pageview in descending order}
\usage{
creat_page_table(page_summary)
}
\arguments{
\item{page_summary}{}
}
\value{
page_df
}
\description{
Create a page summary table to show page (name, clickable hyperlink) and its 
   page view sorted by pageview in descending order
}
\examples{
creat_page_table(page_summary = reactive_page())
}
