% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_server.R
\name{expand_segments}
\alias{expand_segments}
\title{For a given video with a maximum segment, expand to create a dataframe
  with one row per segment. This is a recursive function.}
\usage{
expand_segments(video_id, latest_segment)
}
\arguments{
\item{latest_segment}{The latest segment value. Typically begins with the 
greatest segment value associated with the video in question.}

\item{video_d}{The video ID.}
}
\value{
A dataframe with one row per segment in the video.
}
\description{
For a given video with a maximum segment, expand to create a dataframe
  with one row per segment. This is a recursive function.
}
\examples{
expand_segments("lkjsdflkj1233113lkj23", 6)
}
