% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem_assessment.R
\name{extract_assessment_csv}
\alias{extract_assessment_csv}
\title{Convert a CSV respresenting an open_assessment.sql query into a usable format}
\usage{
extract_assessment_csv(assessment_tbl)
}
\arguments{
\item{assessment_tbl}{}
}
\value{
An extracted table
}
\description{
Extracts the nececessary information from event JSON and discards it. Points
possible should always be the same for each assessment.
}
