% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem_assessment.R
\name{extract_assessment_json}
\alias{extract_assessment_json}
\title{Convert a JSON object into a tidyJSON dataframe}
\usage{
extract_assessment_json(assessment_json)
}
\arguments{
\item{assessment_json}{An JSON file generated through 
`xml_extraction <course> --assessments`}
}
\value{
A flattened dataframe with the columns `url_name`, `title`, `label`, 
  `name`
}
\description{
Convert a JSON object into a tidyJSON dataframe
}
