% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_engagement_server.R
\name{filter_chapter_overview}
\alias{filter_chapter_overview}
\title{Filter course items dataframe by the selected course module}
\usage{
filter_chapter_overview(input_df, module = "All")
}
\arguments{
\item{module}{One of the modules of the course}

\item{item_df}{The course items dataframe.}
}
\value{
A dataframe filtered by the selected course module
}
\description{
Filter course items dataframe by the selected course module
}
\examples{
filter_chapter_oveview(tower_item, "all")
}
