% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem_module.R
\name{filter_counts}
\alias{filter_counts}
\title{Count the number of filtered users for each problem}
\usage{
filter_counts(joined_user_problems)
}
\arguments{
\item{joined_user_problems}{The result of `join_users_problems`}
}
\value{
A dataframe with the count of unique users for each question
}
\description{
Count the number of filtered users for each problem
}
