% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_engagement_server.R
\name{filter_demographics}
\alias{filter_demographics}
\title{Filter input dataframe based on selected student demographics,}
\usage{
filter_demographics(engage_df, gender = "All", mode = "All",
  activity_level = "All")
}
\arguments{
\item{engage_df}{The student engagement dataframe.}

\item{gender}{One of "male", "female", "other", or (default) "all".}

\item{activity_level}{It refers to the total time each student spent on the course.There are 4 levels: "under_30_min", "30_min_to_5_hr", "over_5_hr", or (default) "all".}

\item{registration_status}{One of "audit", "verified", or (default) "all".}

\item{module}{One of the modules of the course.}
}
\value{
A dataframe filtered by the selected demographics.
}
\description{
Filter input dataframe based on selected student demographics,
}
\examples{
filter_demographics(engage_df = tower_engage, gender = "all", mode = "all", activity_level = "all")
}
