% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem_module.R
\name{filter_extreme_problem_choices}
\alias{filter_extreme_problem_choices}
\title{Retrieve the question and choice information from the extreme problems}
\usage{
filter_extreme_problem_choices(lookup_table, extreme_problems, filtered_counts)
}
\arguments{
\item{lookup_table}{The result of `create_question_lookup_from_json`}

\item{extreme_problems}{The result of `get_extreme_summarised_score`}

\item{filtered_counts}{The result of `filter_counts`}
}
\value{
A dataframe with lookup information attached
}
\description{
Retrieve the question and choice information from the extreme problems
}
