% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem_module.R
\name{filter_valid_questions}
\alias{filter_valid_questions}
\title{Filter out invalid multiple choice problems}
\usage{
filter_valid_questions(problems_tbl, lookup_table)
}
\arguments{
\item{problems_tbl}{The result of `read_multiple_choice_csv`}

\item{lookup_table}{The result of `create_question_lookup_from_json`}
}
\value{
The filtered problems_tbl
}
\description{
Removes problems that were marked as 100% correct (typically surveys or
problems where every answer is correct). Also checks from the dataframe
derived from xbundle to ensure that the problem exists in the course.
}
