% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forum_server.R
\name{filter_wordcloud_data}
\alias{filter_wordcloud_data}
\title{Filter the wordcloud data by the selected filters.}
\usage{
filter_wordcloud_data(forum_words = wrangled_forum_words, activity_level,
  gender, registration_status, category)
}
\arguments{
\item{forum_words}{The forum words dataframe passed in from the wrangling script.}

\item{activity_level}{The activity level of the students.}

\item{gender}{The gender of the students.}

\item{registration_status}{The registration status of the students.}

\item{category}{The forum category.}
}
\value{
\code{filtered_wordcloud_data} The filtered forum words dataframe.
}
\details{

}
\examples{
filter_wordcloud_data(wrangled_forum_words, "30_min_to_5_hr", "female", "All", "General")
}
