% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forum_server.R
\name{gather_post_types}
\alias{gather_post_types}
\title{Gather the post types into a single column for easy plotting.}
\usage{
gather_post_types(forum_data, grouping_var)
}
\arguments{
\item{forum_data}{The forum data for the main barplot.}

\item{grouping_var}{grouping_var One of \code{discussion_category} or \code{display_name}, depending on whether a category has been selected.}
}
\value{
\code{gathered} A dataframe with the post types gathered into a single column.
}
\details{

}
\examples{
gather_post_types(forum_data, grouping_var = "discussion_category")
}
