% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forum_wrangling.R
\name{get_activity_levels}
\alias{get_activity_levels}
\title{Convert sum_dt to activity_level.}
\usage{
get_activity_levels(forum)
}
\arguments{
\item{forum}{A forum dataframe.}
}
\value{
The same dataframe with a new column for activity_level.
}
\description{
Convert sum_dt to activity_level.
}
\examples{
get_activity_levels(forum)
}
