% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_server.R
\name{get_age_df}
\alias{get_age_df}
\title{Returns dataframe with student's ages}
\usage{
get_age_df(data)
}
\arguments{
\item{data}{Dataframe containing year of birth}
}
\value{
Returns a dataframe with student's ages
}
\description{
Returns dataframe with student's ages
}
