% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_server.R
\name{get_age_plot}
\alias{get_age_plot}
\title{Returns histogram plot of student's age}
\usage{
get_age_plot(age_df)
}
\arguments{
\item{age_df}{Data frame containing age of students}
}
\value{
Returns a ggplot geom_histogram object
}
\description{
Returns histogram plot of student's age
}
