% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_server.R
\name{get_avg_time_spent}
\alias{get_avg_time_spent}
\title{Returns a dataframe with the average time learners spent on each video}
\usage{
get_avg_time_spent(filtered_segments)
}
\arguments{
\item{filtered_segments}{Dataframe of segments and corresponding watch counts 
filtered by demographics}
}
\value{
\code{g}: A dataframe with the average time learners spent on each 
  video
}
\description{
Returns a dataframe with the average time learners spent on each video
}
\examples{
get_avg_time_spent(filtered_segments)
}
