% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_server.R
\name{get_ch_markers}
\alias{get_ch_markers}
\title{Obtains locations of chapter lines to be placed on visualizations}
\usage{
get_ch_markers(filt_segs)
}
\arguments{
\item{filt_segs}{Dataframe containing students that have been filtered by selected demographics. Typically
obtained via \code{filter_demographics()}}
}
\value{
\code{ch_markers}: List of values of where to place chapter lines on visualizations
}
\description{
Obtains locations of chapter lines to be placed on visualizations
}
\examples{
get_ch_markers(filt_segs)
}
