% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_server.R
\name{get_country_plot}
\alias{get_country_plot}
\title{Returns a plot containing student country distribution}
\usage{
get_country_plot(country_df, top_selection)
}
\arguments{
\item{country_df}{Data frame of students aggregated by country}

\item{top_selection}{The number of countries to include in the plot}
}
\value{
Returns a ggplot geom_bar object
}
\description{
Returns a plot containing student country distribution
}
