% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_wrangling.R
\name{get_end_time}
\alias{get_end_time}
\title{Calculates video end time for non-video events using time stamps}
\usage{
get_end_time(start, time, time_ahead, latest_speed)
}
\arguments{
\item{start}{Time into video that the user has started watching the video}

\item{time}{Time stamp of when the user started watching the video}

\item{time_ahead}{Time stamp of next event following the play event}

\item{latest_speed}{The speed at which the user was watching the video}
}
\value{
\code{end}: Time into video that the user has stopped watching
}
\description{
Calculates video end time for non-video events using time stamps
}
\examples{
get_end_time(start, time, time_ahead, latest_speed)
}
