% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem_module.R
\name{get_extreme_summarised_scores}
\alias{get_extreme_summarised_scores}
\title{Select the easiest (or hardest) problems}
\usage{
get_extreme_summarised_scores(summarised_scores, index)
}
\arguments{
\item{summarised_scores}{the result of `get_mean_scores_filterable`}

\item{index}{the number of questions you wish to view}
}
\value{
The summarised scores dataframe with the with the number of rows
  equal to the absolute value of the index.
}
\description{
Set the index negative to receive the hardest problems.
}
