% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forum_server.R
\name{get_forum_threads}
\alias{get_forum_threads}
\title{Get the forum threads for the authors who match the filter settings.}
\usage{
get_forum_threads(forum_posts = wrangled_forum_posts, activity_level, gender,
  registration_status, category, selected_subcategory)
}
\arguments{
\item{forum_posts}{The forum posts dataframe that was passed in from the wrangling script.}

\item{activity_level}{The activity level of the students.}

\item{gender}{The gender of the students.}

\item{registration_status}{The registration status of the students.}

\item{category}{The selected forum category.}

\item{selected_subcategory}{The selected forum subcategory.}
}
\value{
\code{forum_threads} The forum threads that were authored by students for which the filter settings apply.
}
\details{

}
\examples{
get_forum_threads(forum_posts = wrangled_forum_posts, "over_5_hr", "female", "audit", "All", "All")
}
