% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_checksum.R
\name{get_hashed_courses}
\alias{get_hashed_courses}
\title{Writes csv with course short names and their corresponding checksums to the data directory}
\usage{
get_hashed_courses()
}
\value{
No value is returned
}
\description{
Writes csv with course short names and their corresponding checksums to the data directory
}
\examples{
get_hashed_courses()
}
