% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_checksum.R
\name{get_hashed_dashboard}
\alias{get_hashed_dashboard}
\title{Writes csv with dashboard name and their corresponding checksums to the data directory}
\usage{
get_hashed_dashboard()
}
\value{
No value is returned
}
\description{
Writes csv with dashboard name and their corresponding checksums to the data directory
}
\examples{
get_hashed_dashboard()
}
