% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forum_server.R
\name{get_label_lengths}
\alias{get_label_lengths}
\title{Get the lengths of the labels on the main barplot (either the categories or the subcategories).}
\usage{
get_label_lengths(forum_data, category)
}
\arguments{
\item{forum_data}{The forum data for the main barplot.}

\item{category}{The selected category.}
}
\value{
\code{label_lengths} A list with the lengths (in characters) of each label on the main plot.
}
\description{
Get the lengths of the labels on the main barplot (either the categories or the subcategories).
}
\examples{
get_label_lengths(forum_data, "Technical Questions")
}
