% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_server.R
\name{get_language_plot}
\alias{get_language_plot}
\title{Returns plot of language aggregated information}
\usage{
get_language_plot(language_df, top_selection)
}
\arguments{
\item{language_df}{Data frame with aggregated language information}

\item{top_selection}{Number of different languages to include within the plot}
}
\value{
Returns a plot languages with the most number of students who speak it
}
\description{
Returns plot of language aggregated information
}
