% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_server.R
\name{get_loe_df}
\alias{get_loe_df}
\title{Returns aggregated level of education data frame}
\usage{
get_loe_df(data)
}
\arguments{
\item{data}{Data frame containing explicit level of education of students}
}
\value{
Data frame containing aggregated level of education
}
\description{
Returns aggregated level of education data frame
}
