% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_server.R
\name{get_loe_plot}
\alias{get_loe_plot}
\title{Returns plot with level of education of students}
\usage{
get_loe_plot(loe_df)
}
\arguments{
\item{loe_df}{Data frame containing studentes aggregated by level of education}
}
\value{
Returns a ggplot geom_bar object
}
\description{
Returns plot with level of education of students
}
