% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem_module.R
\name{get_mean_scores}
\alias{get_mean_scores}
\title{Get mean scores.}
\usage{
get_mean_scores(problems_tbl, lookup_table)
}
\arguments{
\item{problems_tbl}{The result of `clean_multiple_choice`}

\item{lookup_table}{The result of `create_question_lookup_from_json`}
}
\value{
A dataframe with each problem, the number of users that attempted
  the problem and the percent score
}
\description{
This function is not affected by filters. If you need a function that is, use
`get_mean_scores_filterable`. Removes scores that do not appear in the
xbundle XML or have a 100% percent average. This was done to remove survey
questions masquerading as multiple choice problems.
}
