% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem_module.R
\name{get_mean_scores_filterable}
\alias{get_mean_scores_filterable}
\title{A filterable version of get_mean_scores}
\usage{
get_mean_scores_filterable(filtered_problems_tbl)
}
\arguments{
\item{filtered_problems_tbl}{The result of `read_multiple_choice_csv` filtered
by demographics and chapter}

\item{name_lookup}{A JSON object contain keys for `id`, `problem`,
`chapter_name`, `chapter_name`, `choices`, `correct_id` and `correct`}
}
\value{
A dataframe where each row is a row and the columns show the
  percentage and absolute number of students that got the question correct.
}
\description{
A filterable version of get_mean_scores
}
