% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_wrangling.R
\name{get_mode}
\alias{get_mode}
\title{Obtain most common value from list}
\usage{
get_mode(x)
}
\arguments{
\item{x}{List containing integer values}
}
\value{
\code{mode}: The most common value within the list
}
\description{
Obtain most common value from list
}
\examples{
get_mode(x=c(0,1,2,2,2,3))
}
