% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_server.R
\name{get_module_options}
\alias{get_module_options}
\title{Appends 'All' to the list of chapter names to be used in the filtering panel}
\usage{
get_module_options(chap_name)
}
\arguments{
\item{chap_name}{A list of chapter names belonging to the course}
}
\value{
The list of chapter names appended with 'All'
}
\description{
Appends 'All' to the list of chapter names to be used in the filtering panel
}
\examples{
get_module_options(c('Chapter 1', 'Chapter 2', 'Chapter 3'))
}
