% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_engagement_server.R
\name{get_nactive}
\alias{get_nactive}
\title{Compute how many students engaged with each course item after filtering student demographic}
\usage{
get_nactive(detail_df)
}
\arguments{
\item{detail_df}{}
}
\value{
summary_df
}
\description{
Compute how many students engaged with each course item after filtering student demographic
}
\examples{
get_nactive(detail_df = tower_engage)
}
