% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forum_wrangling.R
\name{get_post_types}
\alias{get_post_types}
\title{Get the post types.}
\usage{
get_post_types(forum)
}
\arguments{
\item{forum}{A forum dataframe.}
}
\value{
The same dataframe with a new column for post_type.
}
\description{
Get the post types.
}
\examples{
get_post_types(forum)
}
