% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_server.R
\name{get_rank}
\alias{get_rank}
\title{Returns the ranking of a vector x}
\usage{
get_rank(x)
}
\arguments{
\item{x}{A vector of numeric values}
}
\value{
\code{g}: The ranking of the values within x
}
\description{
Returns the ranking of a vector x
}
\examples{
get_rank(c(10, 20, 20, 22, 5))
}
