% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forum_server.R
\name{get_reactive_xvar_string}
\alias{get_reactive_xvar_string}
\title{Get the function to pass into the x value of aes_string() in the main barplot.}
\usage{
get_reactive_xvar_string(category)
}
\arguments{
\item{category}{The selected category.}
}
\value{
\code{reactive_xvar_string} A string that matches the function to call for the x value of aes_string().
}
\description{
Get the function to pass into the x value of aes_string() in the main barplot.
}
\examples{
get_reactive_xvar_string("Part 2")
}
