% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_wrangling.R
\name{get_start_end_df}
\alias{get_start_end_df}
\title{Obtains start and end times for video events}
\usage{
get_start_end_df(data)
}
\arguments{
\item{data}{Dataframe containing tracklog data of students. This is obtained typically
through \code{prepare_video_data()}}
}
\value{
\code{start_end_df}: Original dataframe with \code{start} and \code{end} columns
}
\description{
Parses dataframe and adds columns \code{start} and \code{end} showing
the start and end time that a user watched a video
}
\details{
This will have warning messages that can be ignored. Not sure why
they are popping up. Something to do with get_end_time and it being
evaluated with an entire column of data instead of just one value.
This only occurs once per call of get_end_time. I think its an
initialization issue.
}
\examples{
get_start_end_df(data = data)
}
