% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forum_server.R
\name{get_subcategory_options}
\alias{get_subcategory_options}
\title{Get the options for the subcategory options, given the selected category.}
\usage{
get_subcategory_options(category, filtered_forum_elements)
}
\arguments{
\item{category}{The selected category.}

\item{filtered_forum_elements}{The forum elements dataframe, filtered by the selected category.}
}
\value{
\code{subcategory_options} A list of options for the user to select from.
}
\description{
Get the options for the subcategory options, given the selected category.
}
\examples{
get_subcategory_options("Part 1", filtered_forum_elements)
}
