% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_server.R
\name{get_summary_table}
\alias{get_summary_table}
\title{Obtains video summary table to be used on shiny app}
\usage{
get_summary_table(aggregate_df, vid_lengths)
}
\arguments{
\item{aggregate_df}{Dataframe containing students that have been filtered by selected demographics. Typically
obtained via \code{filter_demographics()}}

\item{vid_lengths}{Dataframe containing video ID's with their associated video lengths. Typically
obtained via \code{get_video_lengths()}}
}
\value{
\code{summary_table}: Dataframe relevant summary statistics for each video.
}
\description{
Obtains video summary table to be used on shiny app
}
\examples{
get_summary_table(filt_segs, vid_lengths)
}
