% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forum_server.R
\name{get_target_word_counts}
\alias{get_target_word_counts}
\title{Get the top words for each subcategory.}
\usage{
get_target_word_counts(input_forum)
}
\arguments{
\item{input_forum}{The input dataframe containing a row for each word in the forum.}
}
\value{
\code{word_counts} A dataframe containing the counts for each word in each subcategory.
}
\description{
Get the top words for each subcategory.
}
\examples{
get_target_word_counts(wrangled_forum_words)
}
