% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_server.R
\name{get_top_country_df}
\alias{get_top_country_df}
\title{Returns country aggregated data frame}
\usage{
get_top_country_df(data, top_selection)
}
\arguments{
\item{data}{Data frame containing students' country}

\item{top_selection}{The number of countries to include in output data frame}
}
\value{
A data frame containing countries with the most number of students
}
\description{
Returns country aggregated data frame
}
