% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_server.R
\name{get_top_hotspots_plot}
\alias{get_top_hotspots_plot}
\title{Obtains heatmap with segments of highest watch rate highlighted}
\usage{
get_top_hotspots_plot(filtered_segments, module, filtered_ch_markers)
}
\arguments{
\item{filtered_segments}{Dataframe of segments and corresponding watch counts filtered by demographics}

\item{module}{String of module (chapter) name to display}

\item{filtered_ch_markers}{List of values containing locations of where to put chapter markers}
}
\value{
\code{g}: ggplot heatmap object
}
\description{
Obtains heatmap with segments of highest watch rate highlighted
}
\examples{
get_top_hotspots_plot(filtered_segments, module, filtered_ch_markers)
}
