% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_server.R
\name{get_top_language_df}
\alias{get_top_language_df}
\title{Returns a language aggregated data frame}
\usage{
get_top_language_df(data, top_selection)
}
\arguments{
\item{data}{Data frame containing student language information}

\item{top_selection}{Number of languages to include within the data frame}
}
\value{
A data frame containing languages with the most number of students who speak it
}
\description{
Returns a language aggregated data frame
}
