% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_dashboard.R
\name{get_unhashed_course}
\alias{get_unhashed_course}
\title{Returns plain text of md5 hashed course}
\usage{
get_unhashed_course(hashed_course, course_checksum_df)
}
\arguments{
\item{hashed_course}{String of md5 hash of the requested course short name}
}
\value{
Plain text of requested course
}
\description{
Returns plain text of md5 hashed course
}
\examples{
get_unhashed_dashboard("6c2915bad76b3214647edde0c0d0bdbe")
}
