% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_dashboard.R
\name{get_unhashed_dashboard}
\alias{get_unhashed_dashboard}
\title{Returns plain text of md5 hashed requested dashboard}
\usage{
get_unhashed_dashboard(hashed_dashboard)
}
\arguments{
\item{hashed_dashboard}{String of md5 hash of the requested dashboard}
}
\value{
Plain text of requested dashboard
}
\description{
Returns plain text of md5 hashed requested dashboard
}
\examples{
get_unhashed_dashboard("aa5b8e1625b36fca0bdc6a70bcb8033b")
}
