% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_page_server.R
\name{get_unique_page_name}
\alias{get_unique_page_name}
\title{Remove duplicated name for each page}
\usage{
get_unique_page_name(each_page_df)
}
\arguments{
\item{each_page_df}{}
}
\value{
all_pages
}
\description{
Remove duplicated name for each page
}
\examples{
get_unique_page_name(each_page_df = each_page)
}
