% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_server.R
\name{get_video_lengths}
\alias{get_video_lengths}
\title{Obtains dataframe with length of videos}
\usage{
get_video_lengths(filt_segs)
}
\arguments{
\item{filt_segs}{Dataframe containing students that have been filtered by selected demographics. Typically
obtained via \code{filter_demographics()}}
}
\value{
\code{vid_lengths}: Dataframe with the video lengths associated with each video ID.
}
\description{
Obtains dataframe with length of videos
}
\examples{
get_video_lengths(filt_segs)
}
