% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forum_server.R
\name{get_wordcloud_data}
\alias{get_wordcloud_data}
\title{Get the counts of each word in the selected subcategories.}
\usage{
get_wordcloud_data(filtered_wordcloud_data, filtered_forum_elements, category,
  selected_subcategory)
}
\arguments{
\item{filtered_wordcloud_data}{The filtered forum words dataframe.}

\item{filtered_forum_elements}{The filtered forum elements dataframe.}

\item{category}{The selected category.}

\item{selected_subcategory}{The selected subcategory.}
}
\value{
\code{wordcloud_data} A dataframe with the counts of each word in the selected subcategory(ies).
}
\description{
Get the counts of each word in the selected subcategories.
}
\examples{
get_wordcloud_data(filtered_wordcloud_data, filtered_forum_elements, "Part 1", "Part 1 Lecture 1 Discussion")
}
