% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_page_wrangling.R
\name{hash_username}
\alias{hash_username}
\title{Anonyanonymize the username column as hash strings in the input_df}
\usage{
hash_username(input_df)
}
\arguments{
\item{input_df}{A dataframe containing username column}
}
\value{
secure_df
}
\description{
Anonyanonymize the username column as hash strings in the input_df
}
\examples{
hash_username(input_df = link_dat)
}
