% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forum_wrangling.R
\name{infer_post_subcategories}
\alias{infer_post_subcategories}
\title{Infer the subcategories of each post, even if it's a response post or a comment.}
\usage{
infer_post_subcategories(forum)
}
\arguments{
\item{forum}{A forum dataframe.}
}
\value{
The same dataframe with commentable_id filled in for everything.
}
\description{
Infer the subcategories of each post, even if it's a response post or a comment.
}
