% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_engagement_server.R
\name{join_engagement_item}
\alias{join_engagement_item}
\title{Join filtered summary engagement dataframe with filtered item dataframe to match "item " ,"item name"  and "nactive" 
convert all module item nacitve number to a constant to draw seperator line later}
\usage{
join_engagement_item(filtered_engagement, filtered_item)
}
\arguments{
\item{filtered_engagement}{}

\item{filtered_item}{}
}
\value{
tower_df
}
\description{
Join filtered summary engagement dataframe with filtered item dataframe to match "item " ,"item name"  and "nactive" 
convert all module item nacitve number to a constant to draw seperator line later
}
\examples{
join_engagement_item(filtered_engagement = filtered_tower_engage,filtered_item = filtered_tower_item)
}
