% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem_assessment.R
\name{join_extracted_assessment_data}
\alias{join_extracted_assessment_data}
\title{Join the results of extract_assessment_csv and extract_assessment_json}
\usage{
join_extracted_assessment_data(extracted_csv, extracted_json)
}
\arguments{
\item{extracted_csv}{the result of extract_assessment_csv}

\item{extracted_json}{the result of extract_assessment_json}
}
\value{
A joined dataframe of the two incoming dataframes
}
\description{
This function joins the BigQuery and XML data. This is needed to populate the
  BigQuery data with the title and label fields. If the ID of the assessment
  does not occur in the XML, it is removed before preceding.
}
\examples{
join_extracted_assessment(sample_extracted_assessment_tbl, extracted_content)
}
