% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem_module.R
\name{join_problems_to_lookup}
\alias{join_problems_to_lookup}
\title{Left join extracted problems and lookup table}
\usage{
join_problems_to_lookup(extracted_problems, lookup_table)
}
\arguments{
\item{extracted_problems}{the result of `clean_multiple_choice`}

\item{lookup_table}{The result of `create_question_lookup_from_json`}
}
\value{
The left joined dataframe by id and choice_id, filtering out blank problems
}
\description{
Left join extracted problems and lookup table
}
