% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem_module.R
\name{join_summary_lookup}
\alias{join_summary_lookup}
\title{Join the summary and lookup tables}
\usage{
join_summary_lookup(summarised_scores, lookup_table)
}
\arguments{
\item{summarised_scores}{the result of `get_mean_scores_filterable`}

\item{lookup_table}{The result of `create_question_lookup_from_json`}
}
\value{
The inner join of the two dataframes
}
\description{
Join the summary and lookup tables
}
