% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem_module.R
\name{join_users_problems}
\alias{join_users_problems}
\title{The left join of joined problems and summarised scores}
\usage{
join_users_problems(joined_problems, summarised_scores)
}
\arguments{
\item{joined_problems}{`problems_tbl` and `lookup_table` joined by `id` and
`item_response`}

\item{summarised_scores}{the result of `get_mean_scores_filterable`}
}
\value{
A left joined dataframe of joined problems and summarised scores
}
\description{
The left join of joined problems and summarised scores
}
