% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_engagement_server.R
\name{make_engagement_eiffel_tower}
\alias{make_engagement_eiffel_tower}
\title{Make eiffel tower plot: all video/problem course items vs. number of engaged 
  students}
\usage{
make_engagement_eiffel_tower(tower_data)
}
\description{
Make eiffel tower plot: all video/problem course items vs. number of engaged 
  students
}
\examples{
make_engagement_eiffel_tower(tower_data = reactive_tower_df())
}
