% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_wrangling.R
\name{make_tidy_segments}
\alias{make_tidy_segments}
\title{Returns tidy (more useable) version of input dataframe}
\usage{
make_tidy_segments(data)
}
\arguments{
\item{data}{Dataframe containing segment information. This dataframe is typically obtained through
\code{get_watched_segments()}}
}
\value{
\code{data}: Tidy version of input dataframe.
}
\description{
Returns a tidy, more usable, version of the input dataframe. Segment information is
converted into a single column using \code{gather()}
}
\examples{
make_tidy_segments(data = start_end_df)
}
