% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_server.R
\name{make_video_time_plot}
\alias{make_video_time_plot}
\title{Make a heatmap showing the number of unique viewers across time.}
\usage{
make_video_time_plot(filt_events, axis, module, ch_markers)
}
\arguments{
\item{filt_events}{A dataframe containing the video events filtered down
by the user's selected inputs.}

\item{axis}{A dataframe containing information about each video.}

\item{module}{The selected module, as defined by the user.}

\item{ch_markers}{A dataframe containing the locations of each chapter
separator.}
}
\value{
A ggplot object showing the number of unique viewers across time.
}
\description{
Make a heatmap showing the number of unique viewers across time.
}
