% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forum_server.R
\name{make_wordcloud}
\alias{make_wordcloud}
\title{Make the wordcloud.}
\usage{
make_wordcloud(wordcloud_data, max_words = 20, scale = c(3.5, 1))
}
\arguments{
\item{wordcloud_data}{A dataframe showing the counts of the top words in the 
selected subcategory.}

\item{max_words}{The maximum number of words to show in the wordcloud. 
Default is 20.}

\item{scale}{A list giving the range of sizes to for the display of words in 
the wordcloud. Default is c(3.5,1).}
}
\value{
A wordcloud object.
}
\description{
Make the wordcloud.
}
\examples{
make_wordcloud(wordcloud_data, max_words = 30, scale = c(4,1.5))
}
