% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_wrangling.R
\name{obtain_country_info}
\alias{obtain_country_info}
\title{Obtains a data frame with countries and ISO codes of their associated languages}
\usage{
obtain_country_info()
}
\value{
Data frame with countries and their associated ISO language codes
}
\description{
Obtains a data frame with countries and ISO codes of their associated languages
}
\examples{
obtain_country_info()
}
