% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_wrangling.R
\name{obtain_language_info}
\alias{obtain_language_info}
\title{Obtains a data frame with language information}
\usage{
obtain_language_info()
}
\value{
Returns a data frame with languages and their corresponding ISO codes
}
\description{
Reads in a .csv containing languages and their ISO code.
}
\examples{
obtain_language_info()
}
