% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_wrangling.R
\name{obtain_raw_general_data}
\alias{obtain_raw_general_data}
\title{Obtains raw general demographic data}
\usage{
obtain_raw_general_data(input_course)
}
\arguments{
\item{input_course}{String value of input course short name}
}
\value{
Returns a data frame with the raw general demographic data
}
\description{
Obtains raw general demographic data
}
\examples{
obtain_raw_general_data('psyc1')
}
