% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_wrangling.R
\name{obtain_raw_video_data}
\alias{obtain_raw_video_data}
\title{Reads raw uncleaned .csv into a dataframe}
\usage{
obtain_raw_video_data(input_course, testing = FALSE)
}
\arguments{
\item{input_course}{Name of course directory (ex. psyc1, spd1, marketing, etc)}
}
\value{
\code{data}: Dataframe containing raw student track log information
}
\description{
Reads the raw .csv obtained through rbq.py into a dataframe. For documentation on how to use
rbq.py, please see www.temporaryreferencelink.com
}
\examples{
obtain_raw_video_data(input_course = 'psyc1')
}
