% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem_module.R
\name{plot_aggregated_problems}
\alias{plot_aggregated_problems}
\title{Plot aggregated problems}
\usage{
plot_aggregated_problems(agg_melted_problems)
}
\arguments{
\item{agg_melted_problems}{A dataframe with the problem
and choice names as well as the number of students who chose
each option.}
}
\value{
A facetted ggplot bar chart
}
\description{
This funciton is used to plot the top or bottom questions.
}
