% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem_assessment.R
\name{plot_assessment}
\alias{plot_assessment}
\title{Plot the summary assessment data}
\usage{
plot_assessment(summary_assessment)
}
\arguments{
\item{summary_assessment}{the resulting dataframe from `summary_assessment`}
}
\value{
A faceted bar plot
}
\description{
Plot the summary assessment data
}
