% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem_module.R
\name{plot_problem_chapter_summaries}
\alias{plot_problem_chapter_summaries}
\title{Overview plot (Chapter by Course)}
\usage{
plot_problem_chapter_summaries(chapter_summary_tbl)
}
\arguments{
\item{chapter_summary_tbl}{The aggregated data on a per chapter basis}
}
\value{
A ggplot bar chart
}
\description{
Overview plot (Chapter by Course)
}
