% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec.R
\name{populate_course}
\alias{populate_course}
\title{Populate a course from within R.}
\usage{
populate_course()
}
\description{
Populate a course from within R.
}
\details{
This function runs the bash script \code{populate_course.sh}, which is located in exec/.
}
\examples{
populate_course()
}
