% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem_module.R
\name{prepare_filterable_problems}
\alias{prepare_filterable_problems}
\title{Produce a dataframe clean dataframe of all questions}
\usage{
prepare_filterable_problems(summarised_scores, lookup_table)
}
\arguments{
\item{summarised_scores}{the result of `get_mean_scores_filterable`}

\item{lookup_table}{The result of `create_question_lookup_from_json`}
}
\value{
A dataframe with the question and percent correct.
}
\description{
Returns the question name and it's rounded percentage (out of 100). This
function is used to prepare the data table shown at the bottom of the problem
dashboard.
}
