% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forum_wrangling.R
\name{prepare_json}
\alias{prepare_json}
\title{Prepare the JSON file for joining with the XML file.}
\usage{
prepare_json(json)
}
\arguments{
\item{json}{A JSON file containing information about the course elements.}
}
\value{
A dataframe with four columns: \code{commentable_id}, \code{display_name}, \code{discussion_category}, and \code{discussion_target}.
}
\description{
Prepare the JSON file for joining with the XML file.
}
\examples{
prepare_json(json)
}
