% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_page_wrangling.R
\name{prepare_page_name}
\alias{prepare_page_name}
\title{Get non-video and non-problem element name then create a chapter/module column for all these course elements}
\usage{
prepare_page_name(course_axis)
}
\arguments{
\item{course_axis}{dataframe containing all course elements information}
}
\value{
page_name
}
\description{
Get non-video and non-problem element name then create a chapter/module column for all these course elements
}
\examples{
prepare_page_name(course_axis =  read_course_axis)
}
