% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forum_wrangling.R
\name{prepare_posts}
\alias{prepare_posts}
\title{Prepare the forum posts.}
\usage{
prepare_posts(forum)
}
\arguments{
\item{forum}{A dataframe with one post per row.}
}
\value{
The prepared forum posts data.
}
\description{
Prepare the forum posts.
}
