% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forum_wrangling.R
\name{prepare_searches}
\alias{prepare_searches}
\title{Prepare the forum searches dataframe.}
\usage{
prepare_searches(forum)
}
\arguments{
\item{forum}{A dataframe with one search event per row.}
}
\value{
The forum searches data with a new column for activity level.
}
\description{
Prepare the forum searches dataframe.
}
\examples{
prepare_searches(forum)
}
