% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_page_wrangling.R
\name{prepare_tidy_page}
\alias{prepare_tidy_page}
\title{Prepare tidy page tracklog data for making page summary table}
\usage{
prepare_tidy_page(page_data)
}
\arguments{
\item{page_data}{}
}
\value{
log_data
}
\description{
Prepare tidy page tracklog data for making page summary table
}
\examples{
prepare_tidy_page(page_data = read_page_dirt)
}
