% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forum_wrangling.R
\name{prepare_views}
\alias{prepare_views}
\title{Prepare the forum views dataframe.}
\usage{
prepare_views(forum)
}
\arguments{
\item{forum}{A dataframe with one read event per row.}
}
\value{
The prepared forum views data.
}
\description{
Prepare the forum views dataframe.
}
\examples{
prepare_views(forum)
}
