% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forum_wrangling.R
\name{prepare_words}
\alias{prepare_words}
\title{Prepare the forum words.}
\usage{
prepare_words(forum)
}
\arguments{
\item{forum}{A dataframe with one row per post.}
}
\value{
The dataframe with each row containing a word, prepared for joining with the forum elements.
}
\description{
Prepare the forum words.
}
