% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forum_wrangling.R
\name{prepare_xml}
\alias{prepare_xml}
\title{Prepare the XML file for joining with the JSON file.}
\usage{
prepare_xml(xml)
}
\arguments{
\item{xml}{An XML file containing information about the course elements.}
}
\value{
A dataframe with three columns: \code{display_name}, \code{discussion_category}, and \code{discussion_target}.
}
\description{
Prepare the XML file for joining with the JSON file.
}
