% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_page_wrangling.R
\name{read_course_axis}
\alias{read_course_axis}
\title{Import course_axis.csv files based on specified course folder name}
\usage{
read_course_axis(input_course)
}
\arguments{
\item{input_course}{the short name of the course}
}
\value{
course_axis a dataframe containing course_axis information
}
\description{
Import course_axis.csv files based on specified course folder name
}
