% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_course_csv}
\alias{read_course_csv}
\title{Read CSV course data}
\usage{
read_course_csv(course_name, csv_name, data_dir = c("..", "data"))
}
\arguments{
\item{course_name}{string short name of the course}

\item{csv_name}{string of the name of the csv}

\item{data_dir}{character vector of the relative path}
}
\value{
dataframe of read_csv results
}
\description{
Will add the ".csv" if not included in `csv_name`
}
