% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_engagement_wrangling.R
\name{read_dirt_engagement_data}
\alias{read_dirt_engagement_data}
\title{Reads in tower_engage_dirt.csv file}
\usage{
read_dirt_engagement_data(input_course)
}
\arguments{
\item{input_course}{Name of course directory (ex. psyc1, spd1, marketing, etc)}
}
\value{
\code{tower_engage}: Dataframe containing student engagement information with videos and problems
}
\description{
Reads in tower_engage_dirt.csv file
}
\examples{
read_dirt_engagement_data(input_course = 'psyc1')
}
