% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_page_wrangling.R
\name{read_link_dirt}
\alias{read_link_dirt}
\title{Import external_link_dirt.csv files based on specified course folder name}
\usage{
read_link_dirt(input_course)
}
\arguments{
\item{input_course}{}
}
\value{
link_dirt
}
\description{
Import external_link_dirt.csv files based on specified course folder name
}
\examples{
read_link_dirt(input_couse = "psyc1")
}
