% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_page_wrangling.R
\name{read_page_dirt}
\alias{read_page_dirt}
\title{Import page_dirt.csv files from specified course folder name}
\usage{
read_page_dirt(input_course)
}
\value{
page_dirt
}
\description{
Import page_dirt.csv files from specified course folder name
}
\examples{
read_page_dirt(input_couse = "psyc1")
}
