% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forum_server.R
\name{reset_filters}
\alias{reset_filters}
\title{Resets all the user-defined filters to "All".}
\usage{
reset_filters(session)
}
\arguments{
\item{session}{A Shiny session object. See https://shiny.rstudio.com/reference/shiny/latest/session.html}
}
\value{
None.
}
\details{
The function is called when the user presses the "Reset" button.
}
\examples{
reset_filters(session)
}
