% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec.R
\name{run_bash_script}
\alias{run_bash_script}
\title{Run a bash script.}
\usage{
run_bash_script(script_name)
}
\arguments{
\item{script_name}{A string the gives the full filename of a bash script in the current directory.}
}
\value{
None.
}
\description{
Run a bash script.
}
\examples{
run_bash_script("populate_course.sh")
}
