% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_dashboard.R
\name{run_dashboard_server}
\alias{run_dashboard_server}
\title{Returns the correct shiny module to use based on `requested_dashboard`}
\usage{
run_dashboard_server(requested_dashboard)
}
\arguments{
\item{requested_dashboard}{String value of which dashboard to display}
}
\value{
Returns a shiny module to be used as a server
}
\description{
Returns the correct shiny module to use based on `requested_dashboard`
}
\examples{
run_dashboard_server("video")
}
