% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_dashboard.R
\name{run_dashboard_ui}
\alias{run_dashboard_ui}
\title{Returns shiny module UI based on `requested_dashboard`}
\usage{
run_dashboard_ui(requested_dashboard)
}
\arguments{
\item{requested_dashboard}{String value of which dashboard to display}
}
\value{
Returns a shiny module UI
}
\description{
Returns shiny module UI based on `requested_dashboard`
}
\examples{
run_dashboard_ui("forum")
}
