% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_page_wrangling.R
\name{set_activity_level}
\alias{set_activity_level}
\title{Set three levels for students total spending time on course based on tracklog link data}
\usage{
set_activity_level(link_data)
}
\arguments{
\item{link_data}{dataframe containing the information of student click any links in the edx course}
}
\value{
link_data
}
\description{
Set three levels for students total spending time on course based on tracklog link data
}
\examples{
set_activity_level(link_data = read_link_dirt)
}
