% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forum_server.R
\name{set_breakdown}
\alias{set_breakdown}
\title{Determines whether the plot types breakdown will be shown.}
\usage{
set_breakdown(plot_variable, breakdown)
}
\arguments{
\item{plot_variable}{One of (default) \code{posts}, \code{authors}, or \code{views}.}

\item{breakdown}{One of \code{TRUE} or (default) \code{FALSE}.}
}
\value{
\code{TRUE} or \code{FALSE}
}
\details{
This option only comes into play if the plot variable is 'posts'.
}
\examples{
set_breakdown(plot_variable = "posts", breakdown = FALSE)
}
