% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forum_server.R
\name{set_fill_value}
\alias{set_fill_value}
\title{Set the fill value of the bars in the main barplot, given the plot variable.}
\usage{
set_fill_value(plot_variable)
}
\arguments{
\item{plot_variable}{One of (default) \code{posts}, \code{authors}, or \code{views}.}
}
\value{
\code{fill_value} A hex value representing the color for the bar fill.
}
\description{
Set the fill value of the bars in the main barplot, given the plot variable.
}
\examples{
set_fill_value("views")
}
