% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forum_server.R
\name{set_forum_plot_ylabel}
\alias{set_forum_plot_ylabel}
\title{Set the label of the y-axis in the main barplot.}
\usage{
set_forum_plot_ylabel(plot_variable)
}
\arguments{
\item{plot_variable}{One of (default) \code{posts}, \code{authors}, or \code{views}.}
}
\value{
\code{forum_plot_ylabe} The label of the y-axis in the main barplot.
}
\description{
Set the label of the y-axis in the main barplot.
}
\examples{
set_forum_plot_ylabel("authors")
}
