% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forum_server.R
\name{set_forum_threads_title}
\alias{set_forum_threads_title}
\title{Set the title of the forum threads table.}
\usage{
set_forum_threads_title(category, selected_subcategory)
}
\arguments{
\item{category}{The selected category.}

\item{selected_subcategory}{The selected subcategory.}
}
\value{
\code{forum_threads_title} A string for the title of the forum threads table.
}
\description{
Set the title of the forum threads table.
}
\examples{
set_forum_threads_title("Part 3", "All")
}
