% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forum_server.R
\name{set_wordcloud_title}
\alias{set_wordcloud_title}
\title{Set the title of the wordcloud.}
\usage{
set_wordcloud_title(category, selected_subcategory)
}
\arguments{
\item{category}{The selected category.}

\item{selected_subcategory}{The selected subcategory.}
}
\value{
\code{wordcloud_title} A string for the wordcloud title.
}
\details{

}
\examples{
set_wordcloud_title("Part 3", "All")
}
