% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem_assessment.R
\name{summarise_joined_assessment_data}
\alias{summarise_joined_assessment_data}
\title{Summarise assessment data for plotting}
\usage{
summarise_joined_assessment_data(joint_assessment, trunc_length = 20)
}
\arguments{
\item{joint_assessment}{The result of join_extracted_assessment_data}

\item{trunc_length}{The length of that the label should be truncated to. Do not
set to less than 4}
}
\value{
A summarised dataframe of the average scores in each area.
}
\description{
This function checks that the number of points possible does not vary within
title-label groups.
}
\examples{
summarise_joined_assessment_data(sample_join_extracted_assessment_data, 20)
}
