% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem_module.R
\name{tally_correct_answers}
\alias{tally_correct_answers}
\title{Count the number of correct answers for each question.}
\usage{
tally_correct_answers(joined_scores)
}
\arguments{
\item{joined_scores}{`problems_tbl` and `lookup_table` joined by `id` and
`item_response`}
}
\value{
A summarised dataframe of correct counts
}
\description{
Count the number of correct answers for each question.
}
