% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_wrangling.R
\name{wrangle_general}
\alias{wrangle_general}
\title{Wrangles raw general demographics data}
\usage{
wrangle_general(input_course)
}
\arguments{
\item{input_course}{String value of input course short name}
}
\value{
No value is returned
}
\description{
Reads the input course's raw 'generalized_demographics.csv' and outputs a cleaned version as wrangled_demographics.csv
within the same directory
}
\examples{
wrangle_general('psyc1')
}
