% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_page_wrangling.R
\name{wrangle_link_page}
\alias{wrangle_link_page}
\title{Read in three csv files getting from rbq.py : course_axis.csv, 
  external_link_dirt.csv and page_dirt.csv; perform wrangling; 
  export three csv files for building link_page_dashabord: external_link.csv, 
  page.csv and page_name.csv}
\usage{
wrangle_link_page(input_course)
}
\description{
Read in three csv files getting from rbq.py : course_axis.csv, 
  external_link_dirt.csv and page_dirt.csv; perform wrangling; 
  export three csv files for building link_page_dashabord: external_link.csv, 
  page.csv and page_name.csv
}
\examples{
wrangle_link_page(input_cours = "psyc1")
}
