% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_wrangling.R
\name{wrangle_video}
\alias{wrangle_video}
\title{Outputs cleaned video data into corresponding course directory}
\usage{
wrangle_video(input_course, testing = FALSE)
}
\arguments{
\item{input_course}{String of course directory name}
}
\value{
No value returned
}
\description{
This function will automatically read a file named "generalized_video_heat.csv"
from the specified course directory and output a csv named "wrangled_video.csv" in the same
directory
}
\examples{
wrangle_video(input_course = "psyc1")
}
