% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_wrangling.R
\name{write_general_data}
\alias{write_general_data}
\title{Writes cleaned demographic data frame as a .csv to the specified course}
\usage{
write_general_data(input_course, data)
}
\arguments{
\item{input_course}{String value of course short name}

\item{data}{The data frame containing cleaned demographic data}
}
\value{
No value is returned
}
\description{
Writes cleaned demographic data frame as a .csv to the specified course
}
\examples{
write_general_data('psyc1', cleaned_demographic_data)
}
