% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_page_wrangling.R
\name{write_link_clean}
\alias{write_link_clean}
\title{Export the tidy link dataframe}
\usage{
write_link_clean(input_course, cleaned_data)
}
\arguments{
\item{cleaned_data}{tidy link dataframe}
}
\description{
Export the tidy link dataframe
}
\examples{
write_link_clean(input_course = "psyc1",cleaned_data = get_module_of_link)
}
