% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_page_wrangling.R
\name{write_page_clean}
\alias{write_page_clean}
\title{Export the tidy page dataframe}
\usage{
write_page_clean(input_course, cleaned_data)
}
\arguments{
\item{input_course}{}

\item{cleaned_data}{tidy page dataframe}
}
\description{
Export the tidy page dataframe
}
\examples{
write_page_clean(input_course = "psyc1",cleaned_data = prepare_tidy_page)
}
