% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_page_wrangling.R
\name{write_page_name}
\alias{write_page_name}
\title{Export the tidy course axis dataframe containing all non-video and non-problem elements with module/chapter information}
\usage{
write_page_name(input_course, cleaned_data)
}
\arguments{
\item{input_course}{}

\item{cleaned_data}{tidy page_name dataframe}
}
\description{
Export the tidy course axis dataframe containing all non-video and non-problem elements with module/chapter information
}
\examples{
write_page_name(input_course = "psyc1",cleaned_data = prepare_page_name)
}
