% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_wrangling.R
\name{write_wrangled_video_data}
\alias{write_wrangled_video_data}
\title{Outputs cleaned data as csv}
\usage{
write_wrangled_video_data(input_course, cleaned_data, testing = FALSE)
}
\arguments{
\item{input_course}{Name of course directory (ex. psyc1, spd1, marketing, etc)}

\item{cleaned_data}{Dataframe containing cleaned data. This cleaned data is typically obtained through 
\code{make_tidy_segments()}}
}
\value{
No return value
}
\description{
Writes cleaned data as a csv into the course correct directory
}
\examples{
write_wrangled_video_data(input_course = 'psyc1', cleaned_data=start_end_df)
}
