% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_engagement_server.R
\name{create_module_name}
\alias{create_module_name}
\title{Create a new column "chapter_name" for course item dataframe in order to 
 implement module filtering}
\usage{
create_module_name(item_df)
}
\arguments{
\item{item_df}{A course axis dataframe only containing item name}
}
\value{
item_df A coure axis dataframe adding chapter_name column
}
\description{
Create a new column "chapter_name" for course item dataframe in order to 
 implement module filtering
}
\examples{
create_module_name(item_df = tower_item)
}
