% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_page_server.R
\name{filter_chapter_linkpage}
\alias{filter_chapter_linkpage}
\title{Filter course items dataframe by the selected course module module}
\usage{
filter_chapter_linkpage(input_df, module = "All")
}
\arguments{
\item{module}{One of the modules of the course}

\item{item_df}{The course items dataframe.}
}
\value{
A dataframe filtered by the selected course module
}
\description{
Filter course items dataframe by the selected course module module
}
\examples{
filter_chapter(tower_item, "all")
}
