% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_engagement_server.R
\name{get_module_nactive}
\alias{get_module_nactive}
\title{Count how many filtered students engaged with the filtered course module}
\usage{
get_module_nactive(tower_df)
}
\arguments{
\item{tower_df}{}
}
\value{
student_num
}
\description{
Count how many filtered students engaged with the filtered course module
}
\examples{
get_module_nactive(reactive_tower_df())
}
