% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forum_server.R
\name{get_viewer_count}
\alias{get_viewer_count}
\title{Get the number of viewers for which the selected filter settings apply.}
\usage{
get_viewer_count(forum_views, forum_elements, activity_level, gender,
  registration_status, category)
}
\arguments{
\item{forum_views}{The forum views dataframe that was passed in from the 
wrangling script.}

\item{forum_elements}{The forum elements dataframe.}

\item{activity_level}{The activity level of the students.}

\item{gender}{The gender of the students.}

\item{registration_status}{The registration status of the students.}

\item{category}{The forum category.}
}
\value{
\code{author_count} A string showing the number of authors.
}
\description{
Get the number of viewers for which the selected filter settings apply.
}
\examples{
get_viewer_count(
  wrangled_forum_views, 
  forum_elements, 
  "All", 
  "male", 
  "verified", 
  "All"
)
}
