% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_engagement_server.R
\name{join_engagement_item}
\alias{join_engagement_item}
\title{Join filtered summary engagement dataframe with filtered item dataframe to match "item " ,"item name"  and "nactive" 
convert all module item nacitve number to a constant to draw seperator line later}
\usage{
join_engagement_item(filtered_engagement, filtered_item)
}
\arguments{
\item{filtered_engagement}{Filtered engagement dataframe.}

\item{filtered_item}{Filtered course axis dataframe}
}
\value{
tower_df Dataframe containing item name, item category and how many student engaged with it
}
\description{
Join filtered summary engagement dataframe with filtered item dataframe to match "item " ,"item name"  and "nactive" 
convert all module item nacitve number to a constant to draw seperator line later
}
\examples{
join_engagement_item(filtered_engagement = filtered_tower_engage,filtered_item = filtered_tower_item)
}
