% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_wrangling.R
\name{obtain_raw_video_data}
\alias{obtain_raw_video_data}
\title{Reads raw uncleaned .csv into a dataframe}
\usage{
obtain_raw_video_data(input_course, testing = FALSE)
}
\arguments{
\item{input_course}{Name of course directory (ex. psyc1, spd1, etc)}

\item{testing}{For developer use only. Boolean used to indicate to use 
testing data.}
}
\value{
\code{data}: Dataframe containing raw student track log information
}
\description{
Reads the raw generalized_video_heat.csv obtained through rbq.py 
  into a dataframe.
}
\examples{
obtain_raw_video_data(input_course = 'psyc1')
}
