% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_page_wrangling.R
\name{read_course_axis}
\alias{read_course_axis}
\title{Import course_axis.csv files based on specified course folder name}
\usage{
read_course_axis(input_course)
}
\arguments{
\item{input_course}{}
}
\value{
course_axis
}
\description{
Import course_axis.csv files based on specified course folder name
}
\examples{
read_course_axis(input_couse = "psyc1")
}
