% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_engagement_wrangling.R
\name{read_dirt_engagement_data}
\alias{read_dirt_engagement_data}
\title{Import untidy student engagement csv file as an R dataframe object}
\usage{
read_dirt_engagement_data(input_course)
}
\arguments{
\item{input_course}{Name of course directory (ex. psyc1, spd1, marketing, etc)}
}
\value{
tower_engage: Dataframe containing student engagement information with videos and problems
}
\description{
Import untidy student engagement csv file as an R dataframe object
}
\examples{
read_dirt_engagement_data(input_course = 'psyc1')
}
