% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_wrangling.R
\name{wrangle_video}
\alias{wrangle_video}
\title{Generates cleaned video data as a csv within a specified course directory}
\usage{
wrangle_video(input_course, testing = FALSE)
}
\arguments{
\item{input_course}{String of short name of course directory}
}
\value{
No value returned
}
\description{
This function will automatically read files named 'generalized_video_heat.csv'
and 'generalized_video_axis.csv' from the specified course directory and output a csv named
'wrangled_video_heat.csv' in the same directory
}
\examples{
wrangle_video(input_course = 'psyc1')
}
