% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview_engagement_wrangling.R
\name{write_tidy_engagement_data}
\alias{write_tidy_engagement_data}
\title{Write cleaned data as a csv into the specified course directory}
\usage{
write_tidy_engagement_data(input_course, cleaned_data)
}
\arguments{
\item{input_course}{Name of course directory (ex. psyc1, spd1, marketing, etc)}

\item{cleaned_data}{Dataframe containing cleaned data.}
}
\description{
Write cleaned data as a csv into the specified course directory
}
\examples{
write_tidy_engagement_data(input_course = 'psyc1', cleaned_data = clean_engagement_data)
}
